requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.4
Flask==1.1.1
pytest==5.4.2
pandas==1.1.5
botocore== 1.15.32
boto3==1.12.32
s3fs==0.5.1
fsspec==0.8.4
PyMySQL==1.0.2
matplotlib==3.3.4
opencv-python==4.5.2.52
scipy==1.5.3
scikit-learn==0.24.1
tensorflow==2.5.0
numpy==1.19.5

